+++ test/tb_axi_atop_filter.sv
@@ -115,7 +115,7 @@ module tb_axi_atop_filter #(
 
   // AXI Master
   logic mst_done = 1'b0;
-  axi_test::rand_axi_master #(
+  axi_test::axi_rand_master #(
     .AW(AXI_ADDR_WIDTH), .DW(AXI_DATA_WIDTH), .IW(AXI_ID_WIDTH), .UW(AXI_USER_WIDTH),
     .TA(TA), .TT(TT),
     .MAX_READ_TXNS        (AXI_MAX_READ_TXNS),
@@ -142,7 +142,7 @@ module tb_axi_atop_filter #(
   end
 
   // AXI Slave
-  axi_test::rand_axi_slave #(
+  axi_test::axi_rand_slave #(
     .AW(AXI_ADDR_WIDTH), .DW(AXI_DATA_WIDTH), .IW(AXI_ID_WIDTH), .UW(AXI_USER_WIDTH),
     .TA(TA), .TT(TT),
     .AX_MIN_WAIT_CYCLES   (RESP_MIN_WAIT_CYCLES),